requirements.txt
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.1
psycopg2-binary==2.9.9
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10